refactor: move item component to separated folder
