+++ src/components/header/item/item.tsx
@@ -1,0 +1,20 @@
+import { SelectedItem, StyledItem } from './styled';
+
+import { Container } from '../styled';
+import React from 'react';
+
+export default function Item(props: {
+  href: string;
+  content: string;
+  active?: boolean;
+}) {
+  return (
+    <Container>
+      {props.active ? (
+        <SelectedItem href={props.href}>{props.content}</SelectedItem>
+      ) : (
+        <StyledItem href={props.href}>{props.content}</StyledItem>
+      )}
+    </Container>
+  );
+}
